vitalsd: finish the overall code and make it the nurse it should be
Vitalsd is a simple text-based tool that dumps the vital statistics of a running
system to a serial device so that the vitality of the host machine can be
tracked over time. This makes it ideal for use in scenarios where the underlying
cause of a periodic freeze is unknown, and data is rarely available using normal
mechanisms like dmesg and kernel printks.

Things done in this change:

  - Added a vitalsd systemd service
  - Added argument parsing to the main routine
  - Altered the code to output to stdout when a serial port isn't provided
  - Added samplers for the following:
    - vmstat
    - thermal data from /sys
    - cooling device data from /sys
    - Iteration counters
    - Time sampler
  - Removed the MultiSampler class since it was unused.

Change-Id: I462f3b995d80a79f1f0458b7c804cb58fab49e0d